Clean the code

+++ include/Screen.hpp
@@ -17,8 +17,8 @@ public:
     explicit Screen(uint8_t width, uint8_t height, int sda_pin, int sck_pin, int8_t reset_pin = -1);
     bool begin();
     void clear();
-    void writeText(const char* text, int x, int y, int size = 1);
-    void drawBitmap(const uint8_t* bitmap, int x, int y, int width, int height);
+    void writeText(const char* text, short x, short y, uint8_t size = 1);
+    void drawBitmap(const uint8_t* bitmap, short x, short y, short width, short height);
     void scrollRight();
 
 private:
